option to hide community engagement buttons
